bevy_picking::focus
Function update_focus

pub fn update_focus(
    picking_behavior: Query<'_, '_, &PickingBehavior>,
    pointers: Query<'_, '_, &PointerId>,
    under_pointer: EventReader<'_, '_, PointerHits>,
    pointer_input: EventReader<'_, '_, PointerInput>,
    over_map: Local<'_, HashMap<PointerId, BTreeMap<FloatOrd, Vec<(Entity, HitData)>>>>,
    hover_map: ResMut<'_, HoverMap>,
    previous_hover_map: ResMut<'_, PreviousHoverMap>,
)

Coalesces all data from inputs and backends to generate a map of the currently hovered entities. This is the final focusing step to determine which entity the pointer is hovering over.