bevy_picking
Struct PickingBehavior

pub struct PickingBehavior {
    pub should_block_lower: bool,
    pub is_hoverable: bool,
}

An optional component that overrides default picking behavior for an entity, allowing you to make an entity non-hoverable, or allow items below it to be hovered. See the documentation on the fields for more details.

Fields
should_block_lower: bool

Should this entity block entities below it from being picked?

This is useful if you want picking to continue hitting entities below this one. Normally, only the topmost entity under a pointer can be hovered, but this setting allows the pointer to hover multiple entities, from nearest to farthest, stopping as soon as it hits an entity that blocks lower entities.

Note that the word “lower” here refers to entities that have been reported as hit by any picking backend, but are at a lower depth than the current one. This is different from the concept of event bubbling, as it works irrespective of the entity hierarchy.

For example, if a pointer is over a UI element, as well as a 3d mesh, backends will report hits for both of these entities. Additionally, the hits will be sorted by the camera order, so if the UI is drawing on top of the 3d mesh, the UI will be “above” the mesh. When focus is computed, the UI element will be checked first to see if it this field is set to block lower entities. If it does (default), the focus system will stop there, and only the UI element will be marked as hovered. However, if this field is set to false, both the UI element and the mesh will be marked as hovered.

Entities without the PickingBehavior component will block by default.

is_hoverable: bool

If this is set to false and should_block_lower is set to true, this entity will block lower entities from being interacted and at the same time will itself not emit any events.

Note that the word “lower” here refers to entities that have been reported as hit by any picking backend, but are at a lower depth than the current one. This is different from the concept of event bubbling, as it works irrespective of the entity hierarchy.

For example, if a pointer is over a UI element, and this field is set to false, it will not be marked as hovered, and consequently will not emit events nor will any picking components mark it as hovered. This can be combined with the other field Self::should_block_lower, which is orthogonal to this one.

Entities without the PickingBehavior component are hoverable by default.

Implementations

impl PickingBehavior

pub const IGNORE: Self

This entity will not block entities beneath it, nor will it emit events.

If a backend reports this entity as being hit, the picking plugin will completely ignore it.